+++ src/qcommon/q_shared.h
@@ -544,7 +544,7 @@ typedef enum
 
 #ifdef HUNK_DEBUG
 #define Hunk_Alloc(size, preference)              Hunk_AllocDebug(size, preference, # size, __FILE__, __LINE__)
-void *Hunk_AllocDebug(size_t size, ha_pref preference, char *label, char *file, int line);
+void *Hunk_AllocDebug(unsigned int size, ha_pref preference, char *label, char *file, int line);
 #else
 void *Hunk_Alloc(size_t size, ha_pref preference);
 #endif
